cron-job/requirements.txt
flask
requests
jetpack-io>=0.5.0